generated_code/dev_outputs/plan_high_level_system_architecture_design/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
psycopg2-binary
alembic
python-dotenv
passlib[bcrypt]

# For testing
pytest
httpx